fix(test): share the shell environment barrier
Keep the panic-safe test environment lock beside shell dispatch so path-included integration harnesses compile without a crate-root test_support module.

Re-export that single lock through test_support for existing unit tests, and remove one unnecessary path allocation surfaced by the strict Rust 1.97 lint gate.
