+++ crates/tui/src/commands/epic_dispatch_acceptance.rs
@@ -30,7 +30,7 @@ fn create_dispatch_app(tmpdir: &TempDir) -> App {
         memory_path: tmpdir.path().join("memory.md"),
         notes_path: tmpdir.path().join("notes.txt"),
         mcp_config_path: tmpdir.path().join("mcp.json"),
-        ..crate::test_support::test_tui_options(tmpdir.path().to_path_buf())
+        ..crate::test_support::test_tui_options(tmpdir.path())
     };
     App::new(options, &Config::default())
 }
